Refactor trigger rule simulation lib
